Chore: 프로젝트 경로 구조 및 코드 스타일 설정 개선
- 프로젝트 경로명에서 대문자 및 언더스코어 제거하여 일관성 확보
- build.gradle 수정
 - Spotless 및 EditorConfig 플러그인 설정 추가
 - Checkstyle 설정 개선 및 suppression 파일 적용
- 코드 스타일 정리
 - import 순서 정리 및 불필요한 import 제거
 - Spotless, editorconfigFormat을 통한 자동 포맷 적용
- checkstyleMain / checkstyleTest 빌드 성공 확인
