+++ .github/workflows/ci.yml
@@ -1,23 +1,38 @@
-name: CI (Build & Test)
+name: ci
 on:
-  push: { branches: [ main ] }
   pull_request:
+  push:
+    branches: [ "main" ]
 
 jobs:
-  build-test:
+  build:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+
       - uses: actions/setup-java@v4
         with:
-          distribution: 'temurin'
-          java-version: '17'
-          cache: gradle
-      - name: Grant execute permission for gradlew
+          distribution: temurin
+          java-version: 17
+          cache: gradle     # ← Gradle 캐시
+
+      - name: Make gradlew executable (Linux)
         run: chmod +x ./gradlew
-      - name: Build
-        if: hashFiles('gradlew') != ''
-        run: ./gradlew build --no-daemon
-      - name: Build (Maven)
-        if: hashFiles('mvnw') != ''
-        run: ./mvnw -B verify
+
+      - name: Format code (Spotless + EditorConfig)
+        run: ./gradlew spotlessApply editorconfigFormat
+
+      - name: Build (tests included)
+        run: ./gradlew clean build --no-daemon
+
+      # Checkstyle 리포트 업로드(리뷰에 유용)
+      - name: Upload Checkstyle reports
+        if: always()
+        uses: actions/upload-artifact@v4
+        with:
+          name: checkstyle-reports
+          path: build/reports/checkstyle/*.html
+
+      # editorconfig-checker도 유지하고 싶다면
+      - name: editorconfig-check
+        uses: editorconfig-checker/action-editorconfig-checker@v2
