+++ build.gradle
@@ -2,22 +2,40 @@ plugins {
     id 'java'
     id 'org.springframework.boot' version '3.5.4'
     id 'io.spring.dependency-management' version '1.1.7'
-    id 'org.ec4j.editorconfig' version '0.0.3'
+    id 'org.ec4j.editorconfig' version '0.1.0'
     id 'checkstyle'
+    id 'com.diffplug.spotless' version '6.25.0'
 }
 
 checkstyle {
     maxWarnings = 0
-    configFile = file("${rootDir}/config/naver-checkstyle-rules.xml")
-    configProperties = ["suppressionFile": "${rootDir}/config/naver-checkstyle-suppressions.xml"]
     toolVersion = "8.42"
+
+    // 1) 규칙 파일은 '하나만' 사용 → naver만 사용
+    configFile = file("$rootDir/config/naver-checkstyle-rules.xml")
+    // config = resources.text.fromFile("config/checkstyle/checkstyle.xml")
+
+    // 2) suppressions 파일 경로 '추가' (덮어쓰기 아님)
+    configProperties.put("suppressionFile", file("$rootDir/config/naver-checkstyle-suppressions.xml"))
+    configProperties.put("extraSuppressionFile", file("config/checkstyle/suppressions.xml"))
+}
+
+
+spotless {
+    java {
+        googleJavaFormat('1.22.0')
+        removeUnusedImports()
+        importOrder()
+        formatAnnotations()
+        target 'src/**/*.java'
+    }
 }
 
 compileJava.options.encoding = 'UTF-8'
 compileTestJava.options.encoding = 'UTF-8'
 
 editorconfig {
-    excludes = ['build']
+    excludes = ['build', '.gradle']
 }
 
 check.dependsOn editorconfigCheck
@@ -60,4 +78,10 @@ dependencies {
 
 tasks.named('test') {
     useJUnitPlatform()
+}
+
+tasks.named('check') { dependsOn 'spotlessApply' }
+
+springBoot {
+    mainClass = 'mute.backend.BackEndApplication'
 }
