+++ config/checkstyle/suppressions.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<!DOCTYPE suppressions PUBLIC
+        "-//Checkstyle//DTD SuppressionFilter Configuration 1.2//EN"
+        "https://checkstyle.org/dtds/suppressions_1_2.dtd">
+
+<suppressions>
+    <!-- 특정 클래스만 임시로 checkstyle 제외 -->
+    <suppress files=".*TempController\.java" checks=".*"/>
+    <suppress files=".*LegacyService\.java" checks=".*"/>
+</suppressions>
