--- src/main/java/Team_Mute/back_end/BackEndApplication.java
@@ -1,13 +1,0 @@
-package Team_Mute.back_end;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackEndApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackEndApplication.class, args);
-	}
-
-}
