--- src/main/java/Team_Mute/back_end/domain/member/service/test.java
@@ -1,4 +1,0 @@
-package Team_Mute.back_end.domain.member.service;
-
-public class test {
-}
