--- src/main/java/Team_Mute/back_end/domain/space_admin/entity/SpaceTagMap.java
@@ -1,29 +1,0 @@
-package Team_Mute.back_end.domain.space_admin.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "tb_space_tag_map")
-@Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
-public class SpaceTagMap {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "map_id")
-	private Integer id;
-
-	@ManyToOne(fetch = FetchType.LAZY, optional = false)
-	@JoinColumn(name = "space_id", nullable = false, foreignKey = @ForeignKey(name = "fk_space_tag_map_space"))
-	@org.hibernate.annotations.OnDelete(action = org.hibernate.annotations.OnDeleteAction.CASCADE) // (선택)
-	private Space space;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "tag_id")
-	private SpaceTag tag;
-
-	@Column(name = "reg_date")
-	private LocalDateTime regDate;
-}
