+++ src/main/java/mute/backend/BackEndApplication.java
@@ -1,0 +1,12 @@
+package mute.backend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BackEndApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(BackEndApplication.class, args);
+  }
+}
