+++ src/main/java/mute/backend/domain/spaceadmin/dto/SpaceListResponse.java
@@ -1,0 +1,38 @@
+package mute.backend.domain.spaceadmin.dto;
+
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+@JsonPropertyOrder({
+  "spaceId", "spaceName", "regionName", "categoryName", "userId",
+  "spaceCapacity", "spaceLocation", "spaceDescription", "spaceImageUrl", "detailImageUrls",
+  "tagNames", "spaceIsAvailable", "regDate", "updDate"
+})
+public interface SpaceListResponse {
+  Integer getSpaceId();
+
+  String getRegionName();
+
+  String getCategoryName();
+
+  Integer getUserId();
+
+  String getSpaceName();
+
+  Integer getSpaceCapacity();
+
+  String getSpaceLocation();
+
+  String getSpaceDescription();
+
+  String getSpaceImageUrl();
+
+  String[] getDetailImageUrls();
+
+  Boolean getSpaceIsAvailable();
+
+  String[] getTagNames();
+
+  java.time.LocalDateTime getRegDate();
+
+  java.time.LocalDateTime getUpdDate();
+}
