+++ src/main/resources/application.properties
@@ -1,11 +1,8 @@
 spring.application.name=back-end
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
-
 # JPA
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-
 #import application-secret.properties
 spring.config.import=optional:classpath:application-secret.properties
-
 server.port=${PORT:8080}
