--- src/test/java/Team_Mute/back_end/BackEndApplicationTests.java
@@ -1,13 +1,0 @@
-package Team_Mute.back_end;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackEndApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
