+++ src/test/java/mute/backend/BackEndApplicationTests.java
@@ -1,0 +1,11 @@
+package mute.backend;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackEndApplicationTests {
+
+  @Test
+  void contextLoads() {}
+}
